+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMYX711YFNQ2BH633YXBJBYY"
+        "01KMYYBBHVK1CWV5YCY0X78DGF"
     )
     r"""The unique identifier of the data source"""