+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMYX7124KYZY431T9FW6R039"
+        "01KMYYBBJ0JGDHGFX3T29ABNVX"
     )
     r"""The unique identifier of the data source"""